refactor: moved Query Provider to contexts
